requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
et_xmlfile==2.0.0
greenlet==3.1.1
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
panda==0.3.1
pandas==2.2.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-pkcs12==1.25
setuptools==75.6.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
